tests/wpt/web-platform-tests/html/browsers/the-window-object/the-windowproxy-object/test-window-proxy-menubar-manual.html

<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="utf-8">
    <title>Window Proxy menubar visible flag Test</title>
    <link rel="author" title='JuneyoungOh' href="juneyoung85@gmail.com">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <h1>Description</h1>
    <p>WindowProxy.menubar Test</p>

    <h1>Manual Test Steps:</h1>
    <ol>
        <li>Make the menubar visible in the user agent before executing this test.</li>
        <li>You may need to manually reload afterwards.</li>
    </ol>

    <div id="log"></div>

    <script>
    test(function() {
        assert_not_equals(typeof window.menubar, undefined, 'window.menubar is undefined');
        assert_true(window.menubar.visible);
    }, "window.menubar.visible");
    </script>
</body>
</html>
